patches/requirements.txt
Flask==1.1.2
pymongo==3.9.0
google-cloud-secret-manager==1.0.0